Move the no data results to more a generic directory
