+++ frontend/src/metabase/components/errors/NoRowsError.tsx
