+++ frontend/src/metabase/public/components/PublicNotFound.jsx
@@ -1,8 +1,8 @@
 import cx from "classnames";
 import { t } from "ttag";
 
+import { NoRowsError } from "metabase/components/errors/NoRowsError";
 import CS from "metabase/css/core/index.css";
-import { NoRowsError } from "metabase/query_builder/components/errors/NowRowsError";
 
 import EmbedFrame from "./EmbedFrame";
